+++ R/path.goUpOneDir.R
@@ -1,0 +1,10 @@
+#' Return path up one directory
+#'
+#' @param p path
+#' @return path with last directory removed
+#' @examples
+#' path.goUpOneDir('/Users/test/fish-100')
+#' @export
+path.goUpOneDir <- function(p) {
+  gsub('(.*)(//?[^//]+)$','\\1',p)
+}